Renamed file publisher plugin and updated names and imports
