+++ plugin/publisher/snap-publisher-mock-file/README.md
