+++ plugin/publisher/snap-publisher-mock-file/file/file.go
@@ -36,7 +36,7 @@ import (
 )
 
 const (
-	name       = "file"
+	name       = "mock-file"
 	version    = 3
 	pluginType = plugin.PublisherPluginType
 )
