+++ plugin/publisher/snap-publisher-mock-file/file/file_small_test.go
