+++ plugin/publisher/snap-publisher-mock-file/main_small_test.go
