+++ plugin/publisher/snap-publisher-mock-file/main_test.go
@@ -33,7 +33,7 @@ import (
 )
 
 var (
-	PluginName = "snap-publisher-file"
+	PluginName = "snap-publisher-mock-file"
 	PluginType = "publisher"
 	SnapPath   = os.Getenv("SNAP_PATH")
 	PluginPath = path.Join(SnapPath, "plugin", PluginName)
